requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
requests==2.31.0
googletrans==4.0.0rc1
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
Pillow==10.1.0
Werkzeug==3.0.1
bleach==6.1.0
deep-translator==1.11.4
mistune==3.0.2
psycopg2-binary
